Task 1
A stack can be implemented using a linked list. The first node can serve as the ‘top’ of Stack and ‘push’ and ‘pop’ operations can be implemented by adding and removing nodes at the head of the linked list. Implement the Stack class using a linked list and provide all the standard member functions.