Task 2
Create an organized linked list in ascending order i.e., all the entries should be added in the list in ascending order. In order to do so the insert(int value) member function must be modified in the linked list class. An illustration of the working of the program is presented in the following.
List list ;
list.insert (5) ;	//5
list.insert (2) ;	//2 5
list.insert (7) ;	//2 5 7
list.insert (3) ;	//2 3 5 7